fix: change the bytes attribute of the SupportedKafkaSizeBytesValueItem openapi object
The following attributes in the getInstanceTypesByCloudProviderAndRegion openapi
endpoint are modified and affected:
- ingress_throughput_per_sec
- egress_throughput_per_sec
- max_data_retention_size
- max_message_size

+++ internal/kafka/internal/api/public/model_supported_kafka_size_bytes_value_item.go
@@ -11,5 +11,5 @@ package public
 
 // SupportedKafkaSizeBytesValueItem struct for SupportedKafkaSizeBytesValueItem
 type SupportedKafkaSizeBytesValueItem struct {
-	Bytes float32 `json:"bytes,omitempty"`
+	Bytes int64 `json:"bytes,omitempty"`
 }
